requirements.txt
python-dotenv
pyodbc
azure-identity
azure-mgmt-consumption